Use 'command' instead of '\' to protect commands
Updates the code to use command instead of backslash to escape
commands to protect them from stubs.

I.e. instead of \[ ... ] we use command [ ... ] and instead of
\cat we use command cat and so on.

For reasons I do not understand use of backslash does not actually
protect the commands, in bash anyway, and shunit2 is broken by stubs.
Use of command actually works.