Support running command-line specified list of tests.
This works because `shunit2` assumes/accepts only one argument, which if provided is the name of the file that contains the scripts. This modification takes all additional arguments to be the names of the test functions to be run.

By example, will support either:
    ./test-script.sh -- test1 ... testN
or
    shunit2 test-script.sh test1 ... testN

The first case requires an additional no-op argument ('--') to tell `shunit2` that additional test arguments are coming.

Resolves #23