Enable colouring if tput supports >= 8 colours
Before this, the $SHUNIT_COLOR variable was defaulting to off if tput
colors returned less than 16. On some terminals, however, tput colors
returns 8, and coloured output still looks fine. Therefore, it seems
appropriate that "auto" should result in colouring on these terminals.

This change changes the behaviour so that colouring is enabled if tput
colors returns >= 8.