Fix SHELL variable not being properly set in tests
In commit 3e9578d141313150fd939211543a1d315d2b3b9a the shell used in
`shunit2_misc_test.sh` was changed from `${SHUNIT_SHELL:-sh}` to
`${SHELL:-sh}` because `SHUNIT_SHELL` was not set anywhere.

However `exec ${shell_bin}` does not reset the `SHELL` variable, it
inherits it from the running shell.
This means before that change all misc tests were running with `sh` and
after the change the running shell was used instead of the shell that is
supposed to be tested.

This can be fixed by manually setting the `SHELL` variable.

This will now fail the travis build because the most misc tests don't
actually pass on `zsh` and `sh` fails `testIssue84` (at least on my
machine).